Update styles; Slides now offline ready
There is a current copy of *remark.js* placed in the out/-directory.

The file *index.html* now contains some style definitions for the
use in the *presi.md*. This will give editors some more options on
placing text and images on the same page. For an example use see
[beginning-git](https://github.com/5minds/beginning-git).

Miscellaneous Changes:
1. Changed highlight style to *solarized-dark*.
2. Slide numbering note showing total no. of Pages.