Add some images
The Image for the DTM has been edited and along with them the visio file.

Please note the export module, if you want to make any changes. This will
save a lot of sweat and blood. Just take a look into the *ENTWICKLERTOOLS*
Tab. There are serveral export modules to be found.